Rename classes
To differentiate between classes used in the different Jetpack installation flows
