+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/jetpack/JetpackActivationEligibilityErrorFragment.kt
@@ -12,9 +12,8 @@ import com.woocommerce.android.extensions.navigateSafely
 import com.woocommerce.android.extensions.navigateToHelpScreen
 import com.woocommerce.android.ui.base.BaseFragment
 import com.woocommerce.android.ui.compose.theme.WooThemeWithBackground
-import com.woocommerce.android.ui.jetpack.benefits.JetpackActivationEligibilityErrorViewModel
-import com.woocommerce.android.ui.jetpack.benefits.JetpackActivationEligibilityErrorViewModel.OpenUrlEvent
-import com.woocommerce.android.ui.jetpack.benefits.JetpackActivationEligibilityErrorViewModel.StartJetpackActivationForApplicationPasswords
+import com.woocommerce.android.ui.jetpack.JetpackActivationEligibilityErrorViewModel.OpenUrlEvent
+import com.woocommerce.android.ui.jetpack.JetpackActivationEligibilityErrorViewModel.StartJetpackActivationForApplicationPasswords
 import com.woocommerce.android.ui.main.AppBarStatus
 import com.woocommerce.android.util.ChromeCustomTabUtils
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event.Exit
