+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/jetpack/benefits/JetpackBenefitsDialog.kt
@@ -62,7 +62,7 @@ class JetpackBenefitsDialog : DialogFragment() {
             when (event) {
                 is JetpackBenefitsViewModel.StartJetpackActivationForJetpackCP -> {
                     findNavController().navigateSafely(
-                        JetpackBenefitsDialogDirections.actionJetpackBenefitsDialogToJetpackInstallStartDialog()
+                        JetpackBenefitsDialogDirections.actionJetpackBenefitsDialogToJetpackCPInstallStartDialog()
                     )
                 }
